organize directory. modified DB_user_time_log now uses email and time in as PKEY.
