+++ gzTrackz/AndroidManifest.xml
@@ -5,7 +5,7 @@
     android:versionName="1.0" >
 
     <uses-sdk
-        android:minSdkVersion="8"
+        android:minSdkVersion="11"
         android:targetSdkVersion="18" />
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -28,7 +28,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.example.gztrackz.RegisterActivity"
+            android:name="com.example.gztrackz.Activity_Register"
             android:label="@string/app_name"
             android:theme="@style/dialog"
             android:screenOrientation="portrait" >
