+++ gzTrackz/bin/AndroidManifest.xml
@@ -5,11 +5,11 @@
     android:versionName="1.0" >
 
     <uses-sdk
-        android:minSdkVersion="8"
+        android:minSdkVersion="11"
         android:targetSdkVersion="18" />
 
     <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
         android:allowBackup="true"
@@ -19,26 +19,34 @@
         <activity
             android:name="com.example.gztrackz.MainActivity"
             android:label="@string/app_name"
-            android:theme="@android:style/Theme.NoTitleBar" >
+            android:theme="@android:style/Theme.NoTitleBar"
+            android:screenOrientation="portrait" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
+                <category android:name="android.intent.category.LAUNCHER"
+                     />
             </intent-filter>
         </activity>
         <activity
-            android:name="com.example.gztrackz.RegisterActivity"
+            android:name="com.example.gztrackz.Activity_Register"
             android:label="@string/app_name"
-            android:theme="@style/dialog" >
+            android:theme="@style/dialog"
+            android:screenOrientation="portrait" >
         </activity>
         <activity
             android:name="com.example.gztrackz.TimeManager"
-            android:label="@string/title_activity_time_manager" >
+            android:label="@string/title_activity_time_manager"
+            android:screenOrientation="portrait" >
         </activity>
-         <activity
+        <activity
             android:name="com.example.gztrackz.TabsManager"
-            android:label="@string/title_activity_tabs_manager" 
-            >
+            android:label="@string/title_activity_tabs_manager"
+            android:screenOrientation="portrait" >
+        </activity>
+        <activity
+            android:name="com.example.gztrackz.StandUpsDialog"
+            android:label="@string/title_activity_stand_ups_dialog"
+            android:screenOrientation="portrait" >
         </activity>
     </application>
