+++ gzTrackz/gen/com/example/gztrackz/R.java
@@ -34,9 +34,9 @@ public final class R {
         public static final int roundbuttonstyle=0x7f020007;
     }
     public static final class id {
-        public static final int action_settings=0x7f090015;
+        public static final int action_settings=0x7f090014;
         public static final int confirmpasswordEditText=0x7f090008;
-        public static final int dateTxt=0x7f090014;
+        public static final int dateTxt=0x7f090013;
         public static final int emailRegisterEditText=0x7f090006;
         public static final int firstnameEditText=0x7f09000d;
         public static final int firstnameTextView=0x7f09000c;
@@ -45,7 +45,6 @@ public final class R {
         public static final int lastnameEditText=0x7f09000f;
         public static final int lastnameTextView=0x7f09000e;
         public static final int logInButton=0x7f090005;
-        public static final int logout=0x7f090012;
         public static final int pager=0x7f090011;
         public static final int passwordEditText=0x7f090001;
         public static final int passwordRegisterEditText=0x7f090009;
@@ -54,17 +53,16 @@ public final class R {
         public static final int textView1=0x7f09000b;
         public static final int textView2=0x7f090007;
         public static final int textView3=0x7f09000a;
-        public static final int timeTxt=0x7f090013;
+        public static final int timeTxt=0x7f090012;
         public static final int usernameEditText=0x7f090002;
     }
     public static final class layout {
         public static final int activity_main=0x7f030000;
         public static final int activity_register=0x7f030001;
         public static final int activity_tabs_manager=0x7f030002;
-        public static final int activity_time_manager=0x7f030003;
-        public static final int fragment_home=0x7f030004;
-        public static final int fragment_standups=0x7f030005;
-        public static final int fragment_timestamp=0x7f030006;
+        public static final int fragment_home=0x7f030003;
+        public static final int fragment_standups=0x7f030004;
+        public static final int fragment_timestamp=0x7f030005;
     }
     public static final class menu {
         public static final int main=0x7f080000;
@@ -76,6 +74,7 @@ public final class R {
         public static final int app_name=0x7f050000;
         public static final int hello_world=0x7f050002;
         public static final int title_activity_register=0x7f050003;
+        public static final int title_activity_stand_ups_dialog=0x7f050006;
         public static final int title_activity_tabs_manager=0x7f050005;
         public static final int title_activity_time_manager=0x7f050004;
     }
